server/requirements.txt
fastapi~=0.135.3
uvicorn
openai>=1.0.0
numpy
faiss-cpu
python-dotenv~=1.2.2
psycopg2-binary
python-jose[cryptography]
python-multipart
python-dateutil
# python-uuid
pydantic~=2.12.5
SQLAlchemy~=2.0.48
pgvector